requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.11,<4.0
djangorestframework-simplejwt>=4.0,<5.0
mysqlclient>=2.0,<3.0
